pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "crude-or-rude"
version = "0.1.0"
description = "A market sentiment analyzer for crude oil news using LangGraph and FastMCP"
authors = ["Crude or Rude Team"]
readme = "README.md"
packages = [{include = "crude_or_rude", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langchain-aws = "^0.2.0"
langchain-core = "^0.3.0"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
pydantic = "^2.0.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
boto3 = "^1.35.0"
mcp = "^1.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"

[tool.poetry.scripts]
crude-or-rude = "crude_or_rude.main:main"
crude-or-rude-server = "crude_or_rude.server_main:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88